# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Skillbytes API - PBL Module",
        "description": "REST API for Problem-Based Learning (PBL) Case Management System",
        "contact": {
            "email": "support@skillbytes.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local Development"
        },
        {
            "url": "http://145.79.13.180/api",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/lessons": {
            "get": {
                "tags": [
                    "Lessons"
                ],
                "summary": "Get paginated lessons",
                "description": "Retrieve paginated list of lessons for authenticated user.",
                "operationId": "e1abe40dbd1870d06943517976c11507",
                "parameters": [
                    {
                        "name": "pagesize",
                        "in": "query",
                        "description": "Page size (default: 15)",
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated lessons",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Lessons retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "slug": {
                                                        "type": "string"
                                                    },
                                                    "title": {
                                                        "type": "string"
                                                    },
                                                    "description": {
                                                        "type": "string"
                                                    },
                                                    "level": {
                                                        "type": "integer"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "pagination": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Lessons"
                ],
                "summary": "Create new lesson",
                "description": "Create a new lesson. Only accessible to admin and guru roles.",
                "operationId": "b196bee153377a601911ff0db5a9aa01",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "slug",
                                    "level_id"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "Lesson title",
                                        "type": "string"
                                    },
                                    "slug": {
                                        "description": "Unique slug",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Lesson description",
                                        "type": "string"
                                    },
                                    "level_id": {
                                        "description": "Level ID",
                                        "type": "integer"
                                    },
                                    "content": {
                                        "description": "Lesson content",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Lesson created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Lesson created successfully"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/lessons/all": {
            "get": {
                "tags": [
                    "Lessons"
                ],
                "summary": "Get all lessons without pagination",
                "description": "Retrieve all available lessons for authenticated user without pagination.",
                "operationId": "9b7f43441a93211d748aebd89ee1665a",
                "responses": {
                    "200": {
                        "description": "All lessons",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "All lessons retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/lessons/{slug}": {
            "get": {
                "tags": [
                    "Lessons"
                ],
                "summary": "Get lesson by slug",
                "description": "Retrieve detailed lesson information by slug.",
                "operationId": "0eb3af9adfc3bb46fdf6f8351af5a895",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Lesson slug",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lesson retrieved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Lesson retrieved successfully"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Lesson not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Lessons"
                ],
                "summary": "Update lesson",
                "description": "Update an existing lesson. Use POST for form-data/file uploads. Only accessible to admin and guru roles.",
                "operationId": "69936bffdef69b4a66f54d3bad82c0ff",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Lesson slug",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "Updated title",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Updated description",
                                        "type": "string"
                                    },
                                    "content": {
                                        "description": "Updated content",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Lesson updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Lesson updated successfully"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Lesson not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Lessons"
                ],
                "summary": "Delete lesson",
                "description": "Delete a lesson permanently. Only accessible to admin and guru roles.",
                "operationId": "3ca7a08e5894f12ceac8afc28c99acc3",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Lesson slug",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lesson deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Lesson deleted successfully"
                                        },
                                        "data": {
                                            "type": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Lesson not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/lessons/by-level/{level}": {
            "get": {
                "tags": [
                    "Lessons"
                ],
                "summary": "Get lessons by level",
                "description": "Retrieve all lessons for a specific level.",
                "operationId": "e5bd01cc1dfc7d046d140dd18f62b2c6",
                "parameters": [
                    {
                        "name": "level",
                        "in": "path",
                        "description": "Level number",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lessons by level",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Lessons by level retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Level not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Bearer token for API authentication",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}